Fixed int instead of long id

+++ src/main/java/utn/frd/ejemplo/rest/services/ClienteRest.java
@@ -45,7 +45,7 @@ public class ClienteRest {
     @PUT
     @Consumes({MediaType.APPLICATION_JSON})
     @Path("/{id}")
-    public void edit(@PathParam("id")long id, Cliente cliente){
+    public void edit(@PathParam("id")int id, Cliente cliente){
         ejbClienteFacade.edit(cliente);
     }
     
@@ -53,7 +53,7 @@ public class ClienteRest {
     @DELETE
     @Consumes({MediaType.APPLICATION_JSON, MediaType.TEXT_PLAIN})
     @Path("/{id}")
-    public void remove(@PathParam("id")long id){
+    public void remove(@PathParam("id")int id){
         ejbClienteFacade.remove( ejbClienteFacade.find(id) );
     }
     
@@ -61,7 +61,7 @@ public class ClienteRest {
     @GET
     @Path("/{id}")
     @Produces({MediaType.APPLICATION_JSON})
-    public Cliente findById(@PathParam("id")long id){
+    public Cliente findById(@PathParam("id")int id){
         return ejbClienteFacade.find(id);
     }
 }
